fix(e2e): separate gateway import test and add PR-changed test detection
Split gateway import e2e tests into their own file so they can run
independently with faster setup (only setup_gateway.py instead of all 4
resource scripts). Update the PR e2e workflow to detect changed test
files and include them alongside the strands-bedrock baseline, using
only the main CDK source to reduce CI time.

Constraint: PR workflow must always run strands-bedrock as a baseline
Rejected: Keep gateway in combined suite | setup creates unnecessary resources when running gateway-only
Confidence: high
Scope-risk: narrow
